Track user roles by server. Don't use method_missing to implement permissions methods for User. Move permissions class to a separate file.
